Pandas Profile of Heart Disease Dataset
EDA can be automated using a Python library called Pandas Profiling. It is a great tool to create reports in the interactive HTML format which is quite easy to understand and analyze the data. Let's explore Pandas Profiling to do EDA in a very short time and with just a single line code.